m_sjoin cleanup: no more strcat() calls.
fixed a few bugs lying around.


git-svn-id: svn+ssh://localhost/home/epiphani/home/dalrepo/bahamut/trunk@271 57140981-5af7-11db-9af8-080020d9fdd4

+++ include/h.h
@@ -170,8 +170,8 @@ extern char *find_or_add(char *);
 extern void count_scache(int *, u_long *);
 extern void list_scache(aClient *, aClient *, int, char **);
 
-extern char *MyMalloc(int);
-extern char *MyRealloc(char *, int);
+extern char *MyMalloc(size_t);
+extern char *MyRealloc(char *, size_t);
 
 /*
  * MyFree is defined as a macro in sys.h
